Switch from Meson to CMake.
While some people think CMake is satan's build system, I believe it's
more hackable, and while that doesn't really matter for this project,
I'd prefer if most of my projects use the same build system.

Meson has also been problematic in the past if you use an older version,
particularly with finding dependencies. CMake is a pseudo-standard, so
it should be a better match for what I want to do.